split scheduling test suite into two
